feat: implementacao do padrao DTO para requests e responses blindando a entidade

+++ src/main/java/br/com/techtask/service/TarefaService.java
@@ -1,5 +1,6 @@
 package br.com.techtask.service;
 
+import br.com.techtask.exception.RecursoNaoEncontradoException;
 import br.com.techtask.modelo.Tarefa;
 import br.com.techtask.repositorio.TarefaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,11 +26,19 @@ public class TarefaService {
     }
 
     public Tarefa atualizarTarefa(long id, Tarefa tarefaAtualizada) {
+
+        if (!repository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Erro: Não é possível atualizar. Tarefa de ID " + id + " não existe!");
+        }
         tarefaAtualizada.setId(id);
         return repository.save(tarefaAtualizada);
     }
 
     public void deletarTarefa(long id) {
+
+        if (!repository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Erro: Tarefa de ID " + id + " não encontrado!");
+        }
         repository.deleteById(id);
     }
 }
