+++ src/main/java/br/com/techtask/service/UsuarioService.java
@@ -1,5 +1,6 @@
 package br.com.techtask.service;
 
+import br.com.techtask.exception.RecursoNaoEncontradoException;
 import br.com.techtask.modelo.Usuario;
 import br.com.techtask.repositorio.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,11 +27,18 @@ public class UsuarioService {
     }
 
     public Usuario atualizarUsuario(Long id, Usuario usuario) {
+        if (!repository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Erro: Não é possível atualizar. Usuário de ID " + id + " não existe!");
+        }
         usuario.setId(id);
         return repository.save(usuario);
     }
 
     public void deletarUsuario(Long id) {
+
+        if (!repository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Erro: Usuário de ID " + id + " não encontrado!");
+        }
         repository.deleteById(id);
     }
 }
